requirements.txt
langchain==0.3.4
langchain-openai==0.2.3
python-dotenv==1.0.1
pydantic==2.9.2
